Parallel Black Scholes done

+++ blackscholes.c
@@ -1,5 +1,5 @@
 #include "blackscholes.h"
-double stdDev(long double trials[], long double mean, int M)
+double stdDev(double trials[], double mean, int M)
 {
 	double stddev = 0.0;
 	int i;
@@ -14,8 +14,8 @@ double stdDev(long double trials[], long double mean, int M)
 void blackScholes(double S, double E, double r, double sigma, double T, int M)
 {
 	int i;
-	long double t, mean = 0.0, stddev, confwidth, confmin, confmax;
-	long double trials[M];
+	double t, mean = 0.0, stddev, confwidth, confmin, confmax;
+	double trials[M];
 
 	struct BoxMullerState state;
 	initBoxMullerState(&state);
@@ -41,5 +41,5 @@ void blackScholes(double S, double E, double r, double sigma, double T, int M)
 	printf("sigma \t%lf\n", sigma);
 	printf("T \t%lf\n", T);
 	printf("M \t%d\n", M);
-	printf("Confidence interval: (%Lf, %Lf)\n", confmin, confmax);
+	printf("Confidence interval: (%lf, %lf)\n", confmin, confmax);
 }
